refactor: split up mutations
