+++ src/helper/mutations/revision.ts
@@ -1,0 +1,86 @@
+import { CheckoutRevisionInput, RejectRevisionInput } from '@serlo/api'
+import { gql } from 'graphql-request'
+import { useRouter } from 'next/router'
+import NProgress from 'nprogress'
+
+import { showToastNotice } from '../show-toast-notice'
+import { mutationFetch } from './helper'
+import { useAuthentication } from '@/auth/use-authentication'
+import { useLoggedInData } from '@/contexts/logged-in-data-context'
+
+export type RevisionMutationMode = 'checkout' | 'reject'
+const rejectEntityMutation = gql`
+  mutation rejectRevision($input: RejectRevisionInput!) {
+    entity {
+      rejectRevision(input: $input) {
+        success
+      }
+    }
+  }
+`
+const checkoutEntityMutation = gql`
+  mutation checkoutRevision($input: CheckoutRevisionInput!) {
+    entity {
+      checkoutRevision(input: $input) {
+        success
+      }
+    }
+  }
+`
+const checkoutPageMutation = gql`
+  mutation checkoutPageRevision($input: CheckoutRevisionInput!) {
+    page {
+      checkoutRevision(input: $input) {
+        success
+      }
+    }
+  }
+`
+
+export function useRevisionMutation() {
+  const auth = useAuthentication()
+  const router = useRouter()
+  const loggedInData = useLoggedInData()
+
+  const revisionMutation = async function (
+    mode: RevisionMutationMode,
+    input: RejectRevisionInput,
+    isPage: boolean
+  ) {
+    const isCheckout = mode === 'checkout'
+    const mutation = isPage
+      ? checkoutPageMutation
+      : isCheckout
+      ? checkoutEntityMutation
+      : rejectEntityMutation
+    NProgress.start()
+    const success = await mutationFetch(
+      auth,
+      mutation,
+      input,
+      loggedInData?.strings.mutations.errors
+    )
+
+    if (success) {
+      setTimeout(() => {
+        if (!loggedInData) return
+        showToastNotice(
+          loggedInData.strings.mutations.success[
+            isCheckout ? 'accept' : 'reject'
+          ],
+          'success'
+        )
+        NProgress.done()
+        void router.push(
+          sessionStorage.getItem('previousPathname') || '/entity/unrevised'
+        )
+      }, 100)
+    }
+    return success
+  }
+  return async (
+    mode: RevisionMutationMode,
+    input: RejectRevisionInput | CheckoutRevisionInput,
+    isPage: boolean
+  ) => await revisionMutation(mode, input, isPage)
+}
